pyproject.toml
[tool.poetry]
name = "sncf-cli"
# https://github.com/python-poetry/poetry/issues/140
# => Support automatic versioning setuptools_scm-style?
version = "v0.8.7"
description = "Search SNCF journeys from the CLI."
authors = ["Yann Feunteun, Maxime Mouchet"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "locomotive" }
]
repository = "https://github.com/yafeunteun/sncf-cli"
homepage = "https://github.com/yafeunteun/sncf-cli"
classifiers=[
    "Environment :: Console",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3"
]
keywords = ["sncf", "train", "tgv", "cli"]

[tool.poetry.dependencies]
python = "^3.6"
attrs = "^19.3"
chevron = "^0.13"
click = "^7.0"
dateparser = "^0.7"
faker = "^3.0.0"
geopy = "^1.20"
requests = "^2.20"
tableformatter = "^0.1.4"
text-unidecode = "^1.3"
py-money = "^0.4.0"

[tool.poetry.dev-dependencies]
# https://github.com/python-poetry/poetry/issues/649
# => This is required because there are no release versions of black at all.
black = {version = "^19.10b0", allow-prereleases = true}
coverage = {extras = ["toml"], version = "^5.0.1"}
coveralls = "^1.10.0"
mypy = "^0.761.0"
# https://github.com/PyCQA/pylint/issues/617
# => Search pylint configuration in setup.cfg and pyproject.toml
# => This will be merged in pylint 2.5.0
pylint = { git = "https://github.com/PyCQA/pylint.git", rev = "3fc8c98" }
pytest = "^5.3"
pytest-cov = "^2.8"
pre-commit = "^1.20.0"
sphinx = "^2.3.1"

[tool.poetry.scripts]
sncf-cli = 'locomotive.cli:cli'

[tool.pylint.messages_control]
ignore = "types.py"
disable = "bad-continuation,no-self-use,too-few-public-methods"

[tool.coverage.run]
omit = ["locomotive/data/*"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
